pyproject.toml
[tool.poetry]
name = "train-tracker"
version = "0.1.0"
description = ""
authors = ["George Kaye <georgejkaye@gmail.com>"]
readme = "README.md"
packages = [{include = "train_tracker", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
Jinja2 = "^3.1.2"
termcolor = "^2.3.0"
libsass = "^0.22.0"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.12.11"
types-requests = "^2.31.0.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
types-psycopg2 = "^2.9.21.20240417"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
